Revised interface with ErgoSCF - now added option to use the MOs produced by ErgoSCF, added option to handle spin-polarized and non-polarized calculations. Added other auxiliary functions to deal with ErgoSCF output. Added some additional tests of these functions. Revised the tutorial about step2 calculations with ErgoSCF. Added some minor bug fixed related to Python2 to 3 transition.

+++ src/libra_py/fix_motion.py
@@ -525,7 +525,7 @@ def process_xyz(filename, PT, itime, ftime, verbose=0):
 
     # Determine dimensions
     ndof = R.num_of_rows
-    nat = ndof/3
+    nat = int(ndof/3)
     nsteps = R.num_of_cols
     if verbose>0:
         print("nsteps = ", nsteps)
